switchdev: Add ability to check if netdev is a switch port
When the SWITCHDEV_SYNC notification is introduced we'll need to
generate it for master devices when a slave was added or for the slaves
themselves when they are leaving the master device. However, it only
makes sense for listeners to process this event in case the netdev in
question is a port netdev or constructed on top of one.

Re-implement switchdev_get_lowest_dev() which was recently removed and
returns an underlying port netdev or NULL otherwise.

Signed-off-by: Ido Schimmel <idosch@mellanox.com>
Reviewed-by: Jiri Pirko <jiri@mellanox.com>