switchdev: Rename call_switchdev_notifiers()
Up until now we only had switchdev notification for FDB addition and
deletion, which were sent along with associated info such as MAC
address. However, we are going to add another notification
(SWITCHDEV_SYNC) that doesn't have any info associated with it.

Instead of of adding the 'info' variable in call sites where it's not
really needed, just rename call_switchdev_notifiers() to
call_switchdev_notifiers_info() and later re-add the first, but with the
initialization of 'info' inside the function.

This is consistent with the netdev notification chain that also have
both variants defined.

Signed-off-by: Ido Schimmel <idosch@mellanox.com>
Reviewed-by: Jiri Pirko <jiri@mellanox.com>