switchdev: Generate event to sync with upper layers
A switchdev driver reflects the kernel datapath down to a capable
device, but there situations in which the two can get out of sync.

Currently, the above can happen when master devices (e.g., bridge, team)
are brought into the datapath. Therefore, we need to be able to make
them and every device stacked on top of them replay its configuration
down to the underlying switchdev driver.

Do that by introducing a new switchdev event (SWITCHDEV_SYNC) that is to
be generated whenever a netdev is enslaved to a master device or removed
from one.

In the first case, the event is generated at the end of the enslavement
process (using the master device itself) and also at the end of its
error path (using the slave netdev). The latter is meant to recover any
configuration that might have been lost during the failed enslavement
process. For example, L3 configuration on a port netdev that failed
joining a LAG.

Generate the event at the appropriate places in ndo_{add,del}_slave() of
bridge, team and bond, which are the current master devices supported by
switchdev drivers.

Signed-off-by: Ido Schimmel <idosch@mellanox.com>
Reviewed-by: Jiri Pirko <jiri@mellanox.com>