vlan: Add support for SWITCHDEV_SYNC notifications
In case a SWITCHDEV_SYNC notification was received from a real device
(e.g. bridge, team) we need to make the lower devices aware of the
upper vlan devices.

Do the above by replaying the potentially missing {,PRE}CHANGEUPPER
notifications. Finally, propagate the event for all upper vlan devices
to allow upper layers to replay any missing configuration themselves.

Note that we simply abort on first error and don't perform any rollback,
as this leads to a consistent behavior with regards to these two code
paths:

1) netdev is removed from under a master device
2) netdev failed to be enslaved to a master device due to SWITCHDEV_SYNC
errors (current path)

In both cases the layers on top of the master device aren't aware of the
configuration change, which makes the underlying driver responsible for
its own cleanup.

Signed-off-by: Ido Schimmel <idosch@mellanox.com>