ipv4: Add support for SWITCHDEV_SYNC notifications
switchdev drivers are expected to configure L3 interfaces based on the
presence of an IP address. Therefore, upon SWITCHDEV_SYNC, replay the
configuration of existing IP addresses by generating NETDEV_UP events in
the inetaddr notification chain.

Like in the kernel, FIBs must only be configured on a netdev in case
it's configured as an L3 interface (in_dev is present). Therefore, in
devinet, register the switchdev notification block with an higher
priority than in FIB code.

Signed-off-by: Ido Schimmel <idosch@mellanox.com>
Reviewed-by: Jiri Pirko <jiri@mellanox.com>