ipv4: fib: Add support for SWITCHDEV_SYNC notifications
When SWITCHDEV_SYNC is received for a netdev it means a change occurred
somewhere in the device chain leading to it and that it now might be
capable of offloading routes pointing to it.

Therefore, upon reception of this event, mark any nexthops containing
the netdev as potentially offload-able by setting 'fib_should_offload'
to true.

Next, iterate over all the routing tables and for each table walk its
trie looking for fib aliases using the marked nexthops. If found, send a
FIB_EVENT_ENTRY_ADD notification to let the underlying switchdev driver
know about the route.

Signed-off-by: Ido Schimmel <idosch@mellanox.com>