mlxsw: spectrum: Make netdevice notification block stateful
When SWITCHDEV_SYNC is introduced we'll replay {PRE,}CHANGEUPPER events
to make sure the kernel and hardware datapaths are synced. This means
that the same event can be sent several times.

The current netdevice notification block isn't smart enough to handle
that and will therefore re-apply the configuration each time, which can
lead to multiple references being taken.

Make the block stateful and prevent it from re-applying configuration.

Signed-off-by: Ido Schimmel <idosch@mellanox.com>
Reviewed-by: Jiri Pirko <jiri@mellanox.com>