mlxsw: spectrum: Associate FIDs with a netdev
We currently associate two types of FIDs with a netdev:
1) vFIDs - with the VLAN-unaware bridge they represent
2) rFIDs - with the L3 netdev they represent

But FIDs from the third type (used in the VLAN-aware bridge) are left
left orphan.

When allocating a FID associate it with the VLAN-aware bridge or with a
VLAN device on top of it, in case one is created.

This allows us to have a consistent behavior across all types of FIDs
and will also allow us to make the netdevice notification block stateful
in the next patch.

Signed-off-by: Ido Schimmel <idosch@mellanox.com>
Reviewed-by: Jiri Pirko <jiri@mellanox.com>