mlxsw: spectrum: Configure vPorts according CHANGEUPPER events
Today there are two ops used to configure a device's VLAN filters:

1) The well-known ndo_vlan_rx_{add,kill}_vid()
2) switchdev_port_obj_{add,del}() with PORT_VLAN object

The difference being that the latter also allows to specify whether the
VLAN is PVID and whether it should be untagged on egress. This is
necessary for switch devices that offload bridging functionality.

The mlxsw driver currently implements both, but uses the first set in
order to configure vPorts, which are essentially a representation of
vlan devices.

However, this approach will become problematic when we start to support
device configuration in non-bottom-up approach via SWITCHDEV_SYNC.

A VLAN-aware bridge configured on top of a LAG device will save its VLAN
configuration in 'vlan_info' of the LAG device, which will be propagated
to its slave via vlan_vids_add_by_dev(). The latter will invoke
ndo_vlan_rx_add_vid(), thereby making the driver think a vlan device is
configured on top of it.

Instead, remove these ndos and configure vPorts based on CHANGEUPPER
events sent for vlan devices.

Note that this also means we need to take care of the situation in which
a port is removed from underneath a LAG device that has VLAN devices on
top, as vlan_vids_del_by_dev() essentially becomes a NOP.

Signed-off-by: Ido Schimmel <idosch@mellanox.com>
Reviewed-by: Jiri Pirko <jiri@mellanox.com>