mlxsw: spectrum: Always process NETDEV_UP events
We currently only process NETDEV_UP events in case a RIF isn't already
associated with the reported netdev, but this must be changed in order
to support all configuration orders.

For example, assume team0.10 is configured with an IP address and
sw1p[1-2] are slaves of team0. Upon the enslavement of sw1p3 to team0
the vPort on that port must also be configured to use the same RIF as
the vPorts on sw1p[1-2]. This is forbidden by current code, as the RIF
already exists.

Change the code to always process NETDEV_UP events and rely on checks in
previous commits to guard against multiple references.

Signed-off-by: Ido Schimmel <idosch@mellanox.com>
Reviewed-by: Jiri Pirko <jiri@mellanox.com>