mlxsw: spectrum: Don't wait for netevent to offload neigh
Currently, we wait for NETEVENT_NEIGH_UPDATE and offload the neighbour
if its state is NUD_VALID. However, this never happens if neighbour is
permanent and already exists, thereby causing permanent neighbour misses
in the device.

Instead, program the neighbour into the device's tables during
ndo_neigh_construct(), if its state is valid.

Signed-off-by: Ido Schimmel <idosch@mellanox.com>
Reviewed-by: Jiri Pirko <jiri@mellanox.com>