mlxsw: spectrum: Track member ports in MID entries
A multicast ID (MID) entry represents the group of local ports bound to
an MDB entry.

Since we remove these entries when the reference count of the first
reaches zero, it's important to make sure a single port doesn't take
multiple references on the entry. This will be possible when the Linux
bridge starts to replay MDB entries upon SWITCHDEV_SYNC.

Therefore, we need to be able to track the ports already member in the
MID entry and decide to whether to destroy the entry based on this list
becoming empty.

Track member ports by adding a linked list to the MID struct. The new
per-port struct will be extended in the next patch in order to the track
the MID entries each port is member in, which is useful when port is
removed from bridge.

Signed-off-by: Ido Schimmel <idosch@mellanox.com>