mlxsw: spectrum: Add support for MDB flushing
While we mostly rely on the Linux bridge to tell us when to remove an
MDB entry, there are situations where this isn't possible and we must
do the cleanup ourselves. For example:

1) When a LAG device is used as a bridge port and ports are removed
from the LAG device. The Linux bridge isn't aware of this change and
therefore doesn't perform any cleanup.

Note that in the device an MDB record pointing to a LAG device is
represented as a record pointing to all the individual ports member in
the LAG. Thus, cleanup must be done on a per-port basis.

2) A vlan device is deleted while being bridged. The vlan's
NETDEV_UNREGISTER is received by the bridge after the vlan was detached
from the underlying port netdev(s). switchdev ops can't be propagated to
the driver.

To perform the flushing we need to know in which MDB entries each port /
vPort is member in. This is done by linking the recently added per-port
MID group struct to the port / vPort.

Signed-off-by: Ido Schimmel <idosch@mellanox.com>