bridge: set port fwd mark before linking to upper dev
nbp_switchdev_mark_set is used to set the 'offload_fwd_mark' of
port netdevs so that packets being flooded by the device won't be
flooded twice. The mark assignment happens right after the device
is linked to the upper bridge master device.

Therefore, when a port netdev joins a bridge, listeners to CHANGEUPPER
notifications cannot retrieve the mark assigned to the port yet.

Flip the order of these two operations so that CHANGEUPPER notifications
will happen after the port was assigned an offload_fwd_mark. This way,
listeners to CHANGUPPER will be able to query the current mark and use it
to change other devices' marks.

Signed-off-by: Andy Roulin <aroulin@cumulusnetworks.com>
Suggested-by: Ido Schimmel <idosch@mellanox.com>
Signed-off-by: Ido Schimmel <idosch@mellanox.com>