bridge: add new bridge port API to set/get offload_fwd_mark
nbp_switchdev_mark_set is used to set the 'offload_fwd_mark' of bridge
ports so that packets already forwarded in hardware will not be
forwarded again by the software bridge. Ports belonging to the same
underlying hardware device are assigned the same mark.

nbp_switchdev_mark_set relies on traversing a bridge port's lower
devices until a switch port netdev is reached. The switch port netdev
will give a "switch ID" to be compared when assigning marks to ports.

This traversing method works to recognize that stacked devices, e.g.,
bond interfaces, belong to the same ASIC but it does not work for
tunnel devices such as VxLAN devices. For these tunnel devices, only
the switch drivers really know if they are being offloaded as they
are not linked to switch ports through lower devices.

To solve tunneling case, switch drivers should be able to assign the
same 'offload_fwd_mark' assigned to stacked switch ports devices to
tunnel devices being offloaded.

Add a new bridge API to set/get a bridge port's 'offload_forward_mark'.
Using this API, The switch driver - when (1) a tunnel device joins a
bridge or (2) a switch port joins a bridge that already has a tunnel
device - will get the mark the bridge assigned to the switch ports and
mark the tunnel device with the same mark.

Signed-off-by: Andy Roulin <aroulin@cumulusnetworks.com>
Suggested-by: Ido Schimmel <idosch@mellanox.com>
Signed-off-by: Ido Schimmel <idosch@mellanox.com>