net: Export function to cycle netdev down and up
The VRF driver is using a nice technique to force routes into the
correct routing table when a netdev is linked/unlinked to/from a VRF
master: The netdev is "cycled" so that the networking stack will think
it went down and then up, thereby prompting the IPv4/IPv6 netdev
notifiers to replay host and prefix routes.

The same technique can be used by switch drivers to make sure prefix
routes are programmed correctly into the hardware once they are relevant
for offload.

Export this function so that it can later be used by other modules.

Signed-off-by: Ido Schimmel <idosch@mellanox.com>
Acked-by: Jiri Pirko <jiri@mellanox.com>