mlxsw: spectrum_router: Determine if RIF should be created based on master
Subsequent patches are going to create router interfaces (RIFs) for
netdevs that are unlinked from their master uppers (e.g., bridge/LAG),
in case they have an IP address configured.

When the RIF is created - during NETDEV_CHANGEUPPER - the netdev is no
longer linked to its master upper, but its private flags (e.g.,
'IFF_BRIDGE_PORT') are still set. This is problematic since the driver
does not configure RIFs for bridge/LAG slaves.

Solve this by checking that the netdev is enslaved to a master upper, in
addition to the private flags check.

Signed-off-by: Ido Schimmel <idosch@mellanox.com>