mlxsw: spectrum_router: Expose a function to sync routes and IP addresses
Expose a function that can be invoked on a netdev that has a become a
valid nexthop device for offload. For example, a bridge device with an
IP address to which an mlxsw port was enslaved.

The function configures a router interface (RIF) on the netdev if
IPv4/IPv6 addresses are configured on it. The function also cycles the
netdev so that existing prefix routes will be offloaded to the hardware.

Signed-off-by: Ido Schimmel <idosch@mellanox.com>