devlink: Add packet trap infrastructure
Add the basic packet trap infrastructure that allows device drivers to
register their supported packet traps with devlink.

Each driver is expected to provide basic information about each
supported trap, such as name and ID, but also the supported metadata
types that will accompany each packet trapped via the trap. The
currently supported metadata types are just a timestamp and the input
port, but more will be added in the future. For example, output port and
traffic class.

Beside registering their packet traps with devlink, drivers are also
expected to report trapped packets to devlink along with relevant
metadata. devlink will maintain packets and bytes statistics for each
packet trap and will potentially report the trapped packet with its
metadata to userspace via a new generic netlink multicast group.

The interface towards the drivers is simple and allows devlink to enable
or disable a packet trap. When disabled, it is expected that the driver
will instruct the underlying device to stop sending packets via this
trap to the host CPU.

Signed-off-by: Ido Schimmel <idosch@mellanox.com>