mlxsw: pci: Query and store PCIe bandwidth during init
A subsequent patch will need to set the CPU policer for dropped packets
to 80% of available PCIe bandwidth. Query and store and PCIe bandwidth,
so that it can be used by the switch driver later on.

For I2C bus we simply return 0 as currently there is no use case for
which it is required.

Signed-off-by: Ido Schimmel <idosch@mellanox.com>