ipv6: Notify newly added route if should be offloaded
The function takes care of adding a single route ('struct fib6_info') to
a FIB node. The route in question should only be notified in case it is
added as the first route in the node (lowest metric) or if it is added
as a sibling route to the first route in the node.

The first criterion can be tested by checking if the route is pointed to
by 'fn->leaf'. The second criterion can be tested by checking the new
'notify_sibling_rt' variable that is set when the route is added as a
sibling to the first route in the node.

Signed-off-by: Ido Schimmel <idosch@mellanox.com>