ipv6: Notify multipath route if should be offloaded
In a similar fashion to previous patches, a multipath route should only
be notified in case it is the first route in the node or if it was
appended to such route.

The type of the notification (replace vs. append) is determined based on
the number of routes added ('nhn') and the number of sibling routes. If
the two do not match, then an append notification should be sent.

Signed-off-by: Ido Schimmel <idosch@mellanox.com>