devlink: Add layer 3 generic packet exception traps
Add layer 3 generic packet exception traps that can report trapped packets
and documentation of the traps.

Unlike drop traps, these exception traps also need to inject the packet
to the kernel's receive path. For example, a packet that was
trapped due to host miss need to be injected into the kernel so that it
will trigger an ARP request or a neighbour solicitation message.

Signed-off-by: Amit Cohen <amitc@mellanox.com>