mlxsw: Add new entry type for unreachable packets
Currently, packets that cannot be routed in hardware (e.g., nexthop
device is not upper of mlxsw), are trapped to the kernel for
forwarding.
Such packets are trapped using "RTR_INGRESS0" trap. This trap also
traps packets that hit reject routes (e.g., "unreachable") so that the
kernel will generate the appropriate ICMP error message for them.

Subsequent patch will need to only report to devlink packets that hit a
reject route, which is impossible as long as "RTR_INGRESS0" is
overloaded like that.

Solve this by using "RTR_INGRESS1" trap for packets that hit reject
routes.

Signed-off-by: Amit Cohen <amitc@mellanox.com>