mlxsw: spectrum_router: Protect NVE configuration with router lock
When an NVE tunnel is initialized, the NVE code tries to
promote the host route corresponding to the tunnel's source IP to
perform decapsulation.

Similarly, when a host route is added, the routing code consults the
current NVE configuration to understand if the route should be
programmed to perform decapsulation.

In order to prevent a race between the two code paths, ensure that NVE
tunnel initialization / de-initialization is protected by the router
lock.

Signed-off-by: Ido Schimmel <idosch@mellanox.com>