mlxsw: pci: Query the PCI timeout for gearbox based systems
Until now, the driver had, for all system types, a common hard coded value
for PCI timeout, time to wait for firmware readiness. After introducing new
system types with gearboxes, the firmware boot time were changed since now
it includes gearboxes readiness.

Hence, the driver need to use different values for different system types.
To achieve it, the driver will first wait for the main firmware to be
ready, than will query the timeout it needs to wait until the main firmware
and the gearboxes are ready and than will use the queried value as the
timeout for system readiness.

The driver can differentiate between systems with and without gearboxes by
introducing 2 new system statues:
1. 0x005D - Only the main firmware is ready
2. 0xFFB0 - One or more gearboxes have a fatal error

As for the second system status, 0xFFB0, the driver will fail the probe.

Signed-off-by: Shalom Toledo <shalomt@mellanox.com>
Acked-by: Jiri Pirko <jiri@mellanox.com>
Signed-off-by: Ido Schimmel <idosch@mellanox.com>