mlxsw: spectrum_trap: Avoid using global variables
The trap code currently uses global variables that represent the
supported packet traps and their corresponding listeners. While the
supported traps are identical across all ASICs for now, this is going to
change by subsequent patches.

Therefore, as a preparatory step, store this information in the main
driver structure and have the code fetch it from there.

Signed-off-by: Ido Schimmel <idosch@mellanox.com>