mlxsw: spectrum_trap: Only use valid trap actions
According to the hardware manual, "SET_FW_DEFAULT" trap action should
not be used for "source" packet traps. These are traps whose action is
controlled by the register that configures the trap.

The driver violates this for a few traps:

1. Exceptions: During driver de-initialization the driver sets
   "SET_FW_DEFAULT" when unregistering these traps

2. ACL discards: "SET_FW_DEFAULT" is set when these traps are disabled

Solve this by using "DISCARD" action, instead. In the first case,
"SET_FW_DEFAULT" is meaningless because the driver is teared down
anyway. In the second case, the action is also meaningless because the
disablement of these traps is implemented by applying a policer with a
rate of 1pps and discarding in software the packets that do reach the
CPU.

Signed-off-by: Ido Schimmel <idosch@mellanox.com>
Reviewed-by: Jiri Pirko <jiri@mellanox.com>